Network.API.Pushover

Description

This is an interface to Pushover.net.

Functions interacting with the the Pushover service.

sendMessage :: Message -> IO (Either Response Response)

Send a message.

getReceipt :: Text -> Text -> IO (Either String Receipt)

Get the Receipt for an emergency message.

Message ingredients.

message :: Text -> Text -> Text -> Message

Build a message with defaults given an API Token, User ID, and initial message.

Additional fields for emergency requests.

emergency :: PriorityLevel

An Emergency PriorityLevel with default EmergencyParams.

eparams :: Lens' PriorityLevel (Maybe EmergencyParams)

eparams gets or sets emergency parameters through a PriorityLevel (which may or may not be an Emergency).

retry :: Lens' EmergencyParams Int

expire :: Lens' EmergencyParams Int

callback :: Lens' EmergencyParams Text

tags :: Lens' EmergencyParams Text

data Message

A message to be sent via sendMessage.

Constructors

Message

Fields

_token :: Text

_user :: Text

_body :: Text

_attachment :: ByteString

_device :: Text

_title :: Text

_url :: Text

_urlTitle :: Text

_priority :: PriorityLevel

_sound :: Sound

_timestamp :: Maybe UTCTime

data PriorityLevel

Priority levels for a notification.

Constructors

VeryLow

Low

Normal

High

Emergency EmergencyParams

data Sound

Constructors

None

Bike

Bugle

CashRegister

Classical

Cosmic

Falling

GameLAN

Incoming

Intermission

Magic

Mechanical

PianoBar

Siren

SpaceAlarm

Tugboat

Alien

Climb

Persistent

Echo

UpDown

token :: Lens' Message Text

user :: Lens' Message Text

body :: Lens' Message Text

attachment :: Lens' Message ByteString

device :: Lens' Message Text

title :: Lens' Message Text

url :: Lens' Message Text

urlTitle :: Lens' Message Text

priority :: Lens' Message PriorityLevel

sound :: Lens' Message Sound

Response to sendMessage

data Response

The response to a sendMessage request.

Constructors

Response

Fields

_status :: Int

_request :: Text

_receipt :: Text

_errors :: [Text]

status :: Lens' Response Int

request :: Lens' Response Text

receipt :: Lens' Response Text

errors :: Lens' Response [Text]

data Acknowledgment

An acknowledgment of an emergency message.

Constructors

Acknowledgment

Fields

_ackAt :: UTCTime

_ackBy :: Text

_ackOn :: Text

data Receipt

Receipt represents the current status of receipt of an emergency message.

Constructors

Receipt

Fields

_acknowledged :: Maybe Acknowledgment

_lastDeliveredAt :: Maybe UTCTime

_expired :: Bool

_expiresAt :: Maybe UTCTime

_calledBackAt :: Maybe UTCTime

ackAt :: Lens' Acknowledgment UTCTime

ackBy :: Lens' Acknowledgment Text

ackOn :: Lens' Acknowledgment Text

acknowledged :: Lens' Receipt (Maybe Acknowledgment)

calledBackAt :: Lens' Receipt (Maybe UTCTime)

expired :: Lens' Receipt Bool

expiresAt :: Lens' Receipt (Maybe UTCTime)

lastDeliveredAt :: Lens' Receipt (Maybe UTCTime)

parseResponse :: ByteString -> Either Response Response